bar ready to use
